Add files via upload
The project presents the circuit synthesis of an alarm with incrementing and decrementing timer with the main purpose of displaying the time using digital displays with 7 segments "seven segment display". Nexys4, Artrix-7 FPGA board is used for the hardware implementation of this project. As soon as the FPGA is turned on, the clock starts, and it and the board's own clock generate synchronization signals for the time display, once the start-stop button is activated. The digital clock displays the operating time by default, and the alarm function can be set using the time set that consists of the board's external resources.