Update README.md
The project presents the circuit synthesis of an alarm with an ascending and descending timer, with the main purpose of displaying the time using 7-segment digital displays. The hardware implementation of this project utilizes the Nexys4 board, Artrix-7 FPGA. As soon as the FPGA is powered on, the clock starts, and both the clock and the board's own clock generate synchronization signals for displaying the time when the start-stop button is activated. The digital clock displays the running time by default, and the alarm function can be set using the time set feature, which involves the external resources of the board.